requirements.txt
amqp==5.0.9
arabic-reshaper==2.1.3
asgiref==3.5.0
attrs==21.4.0
aws-xray-sdk==2.9.0
billiard==3.6.4.0
boto3==1.21.3
botocore==1.24.3
Brotli==1.0.9
cachetools==5.0.0
celery==5.2.3
celery-redbeat==2.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
cssselect2==0.5.0
defusedxml==0.7.1
Deprecated==1.2.13
dj-rest-auth==2.2.3
Django==3.2.8
django-allauth==0.48.0
django-appconf==1.0.5
django-cachalot==2.5.0
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-cryptography==1.1
django-debug-toolbar==3.2.4
django-debug-toolbar-template-profiler==2.0.2
django-extensions==3.1.5
django-filter==21.1
django-fsm==2.8.0
django-fsm-admin==1.2.4
django-guardian==2.4.0
django-json-widget==1.1.1
django-modeltranslation==0.18.2
django-mysql==4.5.0
django-redis==5.2.0
django-requests-debug-toolbar==0.0.2
django-restql==0.15.1
django-reversion==5.0.0
django-sequences==2.6
django-timezone-field==4.2.3
djangoql==0.17.1
djangorestframework==3.13.1
djangorestframework-guardian==0.3.0
dnspython==2.2.0
docutils==0.18.1
drf-dynamic-fields==0.3.1
drf-jwt==1.19.2
drf-nested-routers==0.93.4
drf-writable-nested==0.6.3
feedparser==6.0.8
fonttools==4.30.0
future==0.18.2
gevent==21.12.0
google-api-core==2.5.0
google-api-python-client==2.37.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
gql==3.0.0
graphql-core==3.2.0
graphviz==0.19.1
greenlet==1.1.2
html-json-forms==1.1.1
html5lib==1.1
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
jsonschema==4.4.0
kombu==5.2.3
loguru==0.6.0
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.1.0
multidict==6.0.2
mysqlclient==2.1.0
natural-keys==2.0.0
oauthlib==3.2.0
openapi-codec==1.3.2
packaging==21.3
pdfkit==1.0.0
pendulum==2.1.2
Pillow==9.0.1
ply==3.11
postalcodes-ca==0.0.9
prompt-toolkit==3.0.28
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydotplus==2.0.2
pydyf==0.1.2
pyfcm==1.5.4
PyJWT==1.7.1
Pympler==1.0.1
pyparsing==3.0.7
PyPDF3==1.0.6
pyPEG2==2.15.2
pyphen==0.12.0
pyrsistent==0.18.1
python-bidi==0.4.2
python-crontab==2.6.0
python-dateutil==2.8.2
python-pdf==0.39
python3-openid==3.2.0
pytz==2021.3
pytzdata==2020.1
PyVirtualDisplay==3.0
PyYAML==6.0
redis==4.1.4
reportlab==3.6.8
requests==2.27.1
requests-aws4auth==1.1.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.8
s3transfer==0.5.1
sgmllib3k==1.0.0
shortuuid==1.0.8
six==1.16.0
sqlparse==0.4.2
svglib==1.2.1
tblib==1.7.0
tenacity==8.0.1
tinycss2==1.1.1
tqdm==4.63.0
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
weasyprint==54.2
webencodings==0.5.1
win32-setctime==1.1.0
wkhtmltopdf==0.2
wrapt==1.13.3
xhtml2pdf==0.2.6
yarl==1.7.2
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
zopfli==0.2.1
django-pickled-model

